Prove full-corpus L1.1 restoration breakthrough
Crystallize the named RU462K+EN300K field into the 66.11 MiB implicit V7 representation and add operator-aware bounded readout. The fixed 260k proof passes all 13 damage classes above 95%, all lattice gates, clean preservation, false-certainty, RSS, package, and latency budgets without promoting live runtime authority.

+++ src/nanda_wave/lexical_grokking/model.rs
@@ -23,7 +23,7 @@ pub(super) fn surface_script_flags(surface: &str) -> u8 {
 pub(super) struct AtomRecord {
     pub(super) wave_code: AtomWaveCode,
     pub(super) coupling_start: u32,
-    pub(super) coupling_count: u16,
+    pub(super) coupling_count: u32,
     pub(super) support: u16,
 }
